requirements.txt
django==2.2.15 
pytz==2020.1 
sqlparse==0.3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
Pillow==7.2.0
django-filter==2.3.0
PyJWT==1.7.1 
djangorestframework-simplejwt==4.4.0